[CST-6108] refactored password validation exception
