+++ dspace-server-webapp/src/main/java/org/dspace/app/rest/exception/DSpaceApiExceptionControllerAdvice.java
@@ -20,7 +20,6 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.dspace.app.exception.PasswordNotValidException;
 import org.dspace.app.exception.ResourceAlreadyExistsException;
 import org.dspace.app.rest.utils.ContextUtil;
 import org.dspace.authorize.AuthorizeException;
@@ -164,6 +163,7 @@ public class DSpaceApiExceptionControllerAdvice extends ResponseEntityExceptionH
         EPersonNameNotProvidedException.class,
         GroupNameNotProvidedException.class,
         GroupHasPendingWorkflowTasksException.class,
+        PasswordNotValidException.class,
     })
     protected void handleCustomUnprocessableEntityException(HttpServletRequest request, HttpServletResponse response,
                                                             TranslatableException ex) throws IOException {
@@ -189,13 +189,6 @@ public class DSpaceApiExceptionControllerAdvice extends ResponseEntityExceptionH
         sendErrorResponse(request, response, null,
                           "A required parameter is missing",
                           HttpStatus.BAD_REQUEST.value());
-    }
-
-    @ExceptionHandler(PasswordNotValidException.class)
-    protected void handlePasswordNotValidException(HttpServletRequest request, HttpServletResponse resp, Exception ex)
-        throws IOException {
-        String message = "password.error.robust-password";
-        sendErrorResponse(request, resp, null, message, HttpStatus.UNPROCESSABLE_ENTITY.value());
     }
 
     @Override
